+++ merged-packages/smart-transactions-controller/src/index.ts
@@ -20,3 +20,8 @@ export type {
   SmartTransactionStatuses,
   ClientId,
 } from './types';
+export { MetaMetricsEventName, MetaMetricsEventCategory } from './constants';
+export {
+  getSmartTransactionMetricsProperties,
+  getSmartTransactionMetricsSensitiveProperties,
+} from './utils';